yet another way of measuring the time position of a node
now the length of the current node is correctly adjusted if we are in a grace or scaling construct; and lenght of simultanuous sibling nodes is not computed anymore. this means that simply computing the time till the start of the expression is slightly less expensive and quite accurate. and the Item types theirselves define the behaviour. but to get the correct number of beats etc we should improve the event reading in backwards direction further.