dom.py: work around pylint lossage
dom.py with its heavy use of "ly" and "node" is a good test for Python
code checkers. A test that pylint fails. See
https://bitbucket.org/logilab/pylint/issue/466/

There is code in dom.py that we need to check to keep it correct. So work
around the pylint bug - keep "ly" outside the inheritance statements.