Handle Unpitched and DrumNotes
Unpitched and DrumNotes are now separated. But because MusicXML use a meaning of 'unpitched' similar to what we now call drum notes
the naming can remain (with the risk of some confusion of course).

The unpitched note can get properly defined by simply setting the attribute 'pitch'. Thereafter the new note can be created
(but an attribute error has to be catched because Unpitched doesn't have the attribute 'accidental_token').

When preceded by a chord the unpitched note have the same effect as a 'q' with a duration.