clefname2clef: Add missing clefnames
Now the dict with clef names matches LilyPond 2.19.21.
However, when running the export these are not matched because
in these cases the `clefname` argument is None.
So obviously these clefnames are not detected as clefnames
in the first place.
As Frescobaldi doesn't highlight them I assume they are not
known already (probably added to LilyPond later).
So I'm stuck here because I don't know where the clefnames are
parsed in the first place