Modify command handling/behaviour
There are three major updates to the way commands are handled:

1)
Commands never call die(). Instead they raise exceptions. This is
done so they can also be used from ly-server (who wouldn't die but
rather report an error to the client).

2)
Commands *never* write anything directly, neither to stdout/stderr
nor to a file. Instead they are modifying the data structure passed as
argument in-place. Therefore the Option() object is not used in the
commands anymore.
Depending on the type of commands that are processed in the request/
run there are two commands that will implicitly be appended to the
command list: write and write_info.

3)
Instead of the plain cursor a new dict { 'cursor', 'info' } is used in
the commands. Commands update either of them, modifying the cursor or
appending to 'info'. write and write_info use the respective element from
this dict as their data source.