MusicXML export: Properly handle tagline and copyright
Any tagline (including "" and ##f) has been exported to a <tagline />
tag within <identification>. This is invalid MusicXML, and I think it should be generally reconsidered to simply pump unrecognized header items to
score.info[name].

For "tagline" this commit fixes the behaviour by treating a tagline as a
"rights" tag.
According to the specification one or multiple rights tags can be present,
and if more than one is present each must have an (arbitrary) attribute.