XML: Start supporting manual Beam (WIP)
Export "first" beam.
This works only with regular eighth note beams. Anything else won't work.
Only beam nr 1 is created and ended.
But if more than one beam is involved they have to be encoded all.

That means
a) we have to check the number of beams at [ and ]
b) we have to check at each note if it is beamed, and if so
   the current number of beams has to be calculated.
(seems challenging)