XML: Add beam continuation
As described in
http://lists.gnu.org/archive/html/lilypond-user/2018-05/msg00128.html
there is the possibility to simply encode *one* beam and add
<beam>continue</beam> to each enclosed note.

This is no correct MusicXML, but it will correctly be loaded by MuseScore
and Verovio (test at http://www.verovio.org/musicxml.html).

If we agree to stick with that halfway solution (at least for now) I think we
can now (after adding a test file) merge this.