#121: XML: Properly handle \markup
Closes #121
As described in the Issue the previous implementation of \markup
had two flaws:
- direction modifier wasn't respected
- all markups in a bar would be added as a child of the bar
  instead of being attached to the notes

This commit creates each \markup as a <direction> element
immediately preceding the note/rest the markup is attached to.
If there's an explicit direction operator it is respected.
Multiple markups may be attached to a single note, each with its own
direction (or lack thereof).

The handling of Markup() and MarkupElement() objects is already prepared
to dealing with variable formatting, insofar as a Markup()
contains a list of MarkupElement() items, which will later be used to handle
formatting. One \markup will generate one <direction> element. This contains
a sequence of <words> elements (currently only one), which can individually
be assigned formatting attributes.